chore(components): rename `x` prefixed
