+++ browser-extension/app/components/pastbl-pasta.vue
